% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asl_filenames.R
\name{get_asl_filenames}
\alias{get_asl_filenames}
\title{Get ASL Image Filenames}
\usage{
get_asl_filenames(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{get_image_filenames}},
\code{modalities = "ASL"} so it cannot be specified}
}
\value{
Vector fo filenames
}
\description{
Return the filenames for the ASL images
}
\examples{
get_asl_filenames()
}
